+++ src/test/kotlin/blackjack/model/DeckTest.kt
@@ -1,6 +1,5 @@
-package blackjack
+package blackjack.model
 
-import blackjack.model.Deck
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
