+++ src/test/kotlin/blackjack/model/GameManagerTest.kt
@@ -1,6 +1,5 @@
-package blackjack
+package blackjack.model
 
-import blackjack.model.GameManager
 import blackjack.model.participant.Dealer
 import blackjack.model.participant.Player
 import org.assertj.core.api.Assertions.assertThat
