+++ src/test/kotlin/blackjack/model/participant/DealerTest.kt
@@ -1,9 +1,8 @@
-package blackjack.participant
+package blackjack.model.participant
 
 import blackjack.model.card.Card
 import blackjack.model.card.CardNumber
 import blackjack.model.card.Shape
-import blackjack.model.participant.Dealer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
