+++ src/test/kotlin/blackjack/model/participant/PlayerTest.kt
@@ -1,12 +1,10 @@
-package blackjack.participant
+package blackjack.model.participant
 
 import blackjack.model.ResultCalculator
 import blackjack.model.ResultType
 import blackjack.model.card.Card
 import blackjack.model.card.CardNumber
 import blackjack.model.card.Shape
-import blackjack.model.participant.Dealer
-import blackjack.model.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
